style: rename db to database
