+++ app/src/main/java/com/shizq/bika/database/SearchRepository.java
@@ -1,4 +1,4 @@
-package com.shizq.bika.db;
+package com.shizq.bika.database;
 
 import android.content.Context;
 import android.os.AsyncTask;
